Add basic controller tests
Add some basic test for the UsersRestController to ensure they are wired
up to the appropriate service ports (GetUserService and AddUserService).

There is debate about the value of unit testing controllers in spring.
The test here are added as an example of an approach. The value is up to
the individual.

Still need to add tests for the services and perhaps test to the
controller where a request is made to the url endpoint rather than
having the test communicate directly to methods in the controller.