For a user the username can not be null
This validation is performed using Lombok's @NonNull annotation. If the
add user endpoint receives a request that can be transformed into a User
domain model, but has no username, then the request is rejected as a Bad
Request.

The AddUserService never knows about this request.

Of sourse there is a debate over whether this logic should be in a
controller or the domain as the argument can be made that not allowing a
user's name to be null is a business rule and thus is something that is
the responsibility of the domain. However, doing so would prevent the
use of Lombok's NonNull annotation and require more work on our behalf.