Hook the addUserAdapter to the userRepository so users are now 'saved'
We can now save users to the in-memory user repository. However, we have
no way of retrieving the saved users at this commit as the getUser(s) still
returns a hard coded response.

When a user is saved their id is returned back to the requester. We have
also added a business rule /logic that requires that users being saved
must have a name that is not a zero length string.