Can now get a single user by their userId
As each user is added to the repository, they are assigned an id. This
id can be used to get their details. If an id is requested that is not
found within the reposiroty a response with the status of 'not found'
with a message indicating the associated user could not be found is
returned.