Create initial implementation of a name verifier adapter
This adapter serves to 'validate' a name provided. Essentially as long
is it is only made up of alphabetic characters and spaces then it will
be treated as valid.

The reason for breaking this out into its own adapter is to:

Demostrate that the separate adapters have no contact with each other and that the
domain aka the service, handles the processing flow.

That integration with third party apps / apis generally happen in these
adapters away from the domain (just like how we do databases /
repositories). Of course this implemenatation does not really talk
to any third party system, but is more to show that you would be able to
drop in a different adapter module (eg. one that used rest to
communicate out) and the domain would never know.