Domain now uses name-verifier module
When adding a user, the domain verifies the user's name. If the name is
valid the user is saved, otherwise a response of Bad Request is
returned.

The name-verifier-adapter is a fake to simulate the idea of how a third
party app can be called by our example app and isolated to just one
module. Thus the implementation needed to connect to whatever the
name-verifier is is unknown to the domain. This helps to demonstrate the
value of ports and adapters where adapters can easily be changed without
any other part of the codebase / app knowing, caring or needing to be
changed.